requirements.txt
colorama==0.4.6
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pycryptodomex==3.17
pydirectinput-rgx==2.0.7
pyflakes==2.4.0
vgamepad==0.0.8
